Add ability to generate URLs with query strings
