+++ src/Illuminate/Routing/GeneratesQueryStrings.php
@@ -1,0 +1,60 @@
+<?php
+
+namespace Illuminate\Routing;
+
+use Illuminate\Support\Arr;
+
+trait GeneratesQueryStrings
+{
+    /**
+     * Generate a query string with the given parameters.
+     *
+     * @param  array  $parameters
+     * @return string
+     */
+    protected function makeQueryString(array $parameters)
+    {
+        if (count($parameters) === 0) {
+            return '';
+        }
+
+        $query = Arr::query(
+            $keyed = $this->getStringParameters($parameters)
+        );
+
+        // If there happens to be more parameters remaining, we will fetch the numeric
+        // parameters that are in the array and add them to the query string or we
+        // will make the initial query string if it wasn't started with strings.
+        if (count($keyed) < count($parameters)) {
+            $query .= '&'.implode(
+                '&', $this->getNumericParameters($parameters)
+            );
+        }
+
+        $query = trim($query, '&');
+
+        return $query === '' ? '' : "?{$query}";
+    }
+
+    /**
+     * Get the string parameters from a given list.
+     *
+     * @param  array  $parameters
+     * @return array
+     */
+    protected function getStringParameters(array $parameters)
+    {
+        return array_filter($parameters, 'is_string', ARRAY_FILTER_USE_KEY);
+    }
+
+    /**
+     * Get the numeric parameters from a given list.
+     *
+     * @param  array  $parameters
+     * @return array
+     */
+    protected function getNumericParameters(array $parameters)
+    {
+        return array_filter($parameters, 'is_numeric', ARRAY_FILTER_USE_KEY);
+    }
+}
